More work in Kaminari implementation

+++ samples/server/src/kumo/rpc.hpp
@@ -2,6 +2,7 @@
 #include <kumo/opcodes.hpp>
 #include <kumo/rpc_detail.hpp>
 #include <kumo/structs.hpp>
+#include <kumo/protocol_queues.hpp>
 #include <kaminari/buffers/packet.hpp>
 #include <kaminari/broadcaster.hpp>
 class client;
@@ -23,12 +24,12 @@ namespace kumo
 {
     inline void send_do_sth(::kumo::protocol_queues* pq, complex&& data, T&& callback)
     {
-        pq->send_reliable(opcode::do_sth, std::move(data), std::forward<T>(callback));
+        pq->send_reliable(static_cast<uint16_t>(opcode::do_sth), std::move(data), std::forward<T>(callback));
     }
     template <typename B, typename T>
     void broadcast_do_sth(::kaminari::broadcaster<B>* broadcaster, complex&& data, T&& callback)
     {
-        boost::intrusive_ptr<::kaminari::packet> packet = ::kaminari::packet::make(opcode::do_sth, std::forward<T>(callback));
+        boost::intrusive_ptr<::kaminari::packet> packet = ::kaminari::packet::make(static_cast<uint16_t>(opcode::do_sth), std::forward<T>(callback));
         ::kumo::pack(packet, data);
         broadcaster->broadcast([packet](auto pq) {
             pq->send_reliable(packet);
@@ -37,7 +38,7 @@ namespace kumo
     template <typename B, typename T>
     void broadcast_do_sth_single(::kaminari::broadcaster<B>* broadcaster, complex&& data, T&& callback)
     {
-        boost::intrusive_ptr<::kaminari::packet> packet = ::kaminari::packet::make(opcode::do_sth, std::forward<T>(callback));
+        boost::intrusive_ptr<::kaminari::packet> packet = ::kaminari::packet::make(static_cast<uint16_t>(opcode::do_sth), std::forward<T>(callback));
         ::kumo::pack(packet, data);
         broadcaster->broadcast([packet](auto pq) {
             pq->send_reliable(packet);
@@ -45,12 +46,12 @@ namespace kumo
     }
     inline void send_spawn(::kumo::protocol_queues* pq, spawn_data&& data, T&& callback)
     {
-        pq->send_reliable(opcode::spawn, std::move(data), std::forward<T>(callback));
+        pq->send_reliable(static_cast<uint16_t>(opcode::spawn), std::move(data), std::forward<T>(callback));
     }
     template <typename B, typename T>
     void broadcast_spawn(::kaminari::broadcaster<B>* broadcaster, spawn_data&& data, T&& callback)
     {
-        boost::intrusive_ptr<::kaminari::packet> packet = ::kaminari::packet::make(opcode::spawn, std::forward<T>(callback));
+        boost::intrusive_ptr<::kaminari::packet> packet = ::kaminari::packet::make(static_cast<uint16_t>(opcode::spawn), std::forward<T>(callback));
         ::kumo::pack(packet, data);
         broadcaster->broadcast([packet](auto pq) {
             pq->send_reliable(packet);
@@ -59,7 +60,7 @@ namespace kumo
     template <typename B, typename T>
     void broadcast_spawn_single(::kaminari::broadcaster<B>* broadcaster, spawn_data&& data, T&& callback)
     {
-        boost::intrusive_ptr<::kaminari::packet> packet = ::kaminari::packet::make(opcode::spawn, std::forward<T>(callback));
+        boost::intrusive_ptr<::kaminari::packet> packet = ::kaminari::packet::make(static_cast<uint16_t>(opcode::spawn), std::forward<T>(callback));
         ::kumo::pack(packet, data);
         broadcaster->broadcast([packet](auto pq) {
             pq->send_reliable(packet);
